Fix open file dialog stuck
The last commit seems still don't fix the stuck problem, this one
should do the work.
This commit is only a work-around to this problem. But since this
project and deepin-movie both are long time no maintained projects,
I here just introduce this work-around to fix the bug and later
maintainters don't count on this commit please.

Change-Id: I78b035a64d277d706f40b1b313295e91c28d006e